# fmpz_lll_mpf: use simple dot product instead of slow cancellation-checking dot product