--- a/src/fmpz_lll/check_babai_heuristic.c
+++ b/src/fmpz_lll/check_babai_heuristic.c
@@ -56,15 +56,17 @@ fmpz_lll_check_babai_heuristic(int kappa, fmpz_mat_t B, fmpz_mat_t U,
 
             for (j = aa; j < kappa; j++)
             {
-                if (mpf_cmp_d(mpf_mat_entry(A->appSP2, kappa, j), DBL_MIN)
-                    == 0)
+                if (mpf_cmp_d(mpf_mat_entry(A->appSP2, kappa, j), DBL_MIN) == 0)
                 {
-                    if (!
-                        (_mpf_vec_dot2
-                         (mpf_mat_entry(A->appSP2, kappa, j),
-                          appB->rows[kappa], appB->rows[j], n, prec)))
+#if 1
+/* Use slow dot product and check for cancellation */
+                    if (!(_mpf_vec_dot2(mpf_mat_entry(A->appSP2, kappa, j), appB->rows[kappa], appB->rows[j], n, prec)))
+#else
+/* Use fast dot product (always returns 1, pretending that there is no cancellation) */
+                    if (!(_mpf_vec_dot1(mpf_mat_entry(A->appSP2, kappa, j), appB->rows[kappa], appB->rows[j], n, prec)))
+#endif
                     {
-/* In this case a heuristic told us that some cancellation probably happened so we just compute the scalar product at full precision */
+                        /* In this case a heuristic told us that some cancellation probably happened so we just compute the scalar product at full precision */
                         _fmpz_vec_dot(ztmp, fmpz_mat_row(B, kappa), fmpz_mat_row(B, j), n);
                         fmpz_get_mpf(mpf_mat_entry(A->appSP2, kappa, j), ztmp);
                     }