In ``fmpz_lll_check_babai_heuristic``, there is an ``mpf`` dot product computation which attempts to guess if there is too much cancellation and then falls back to an exact ``fmpz`` dot product. Looking at all calls to LLL that occur in our test suite, it seems that this fallback is very rarely useful, while the cancellation check costs more than the actual dot product. We could optimize the cancellation check, but let's try to simplify things and just skip it.

It turns out that this gives a nearly 2x speedup in the polynomial factoring examples which are currently commented out in ``build/fmpz_poly_factor/profile/p-factor_hard`` due to being excessively slow.

Old:

```
H2 has 6 factors: 443192 ms
S9 has 1 factors: 364048 ms
```

New:

```
H2 has 6 factors: 235941 ms
S9 has 1 factors: 201570 ms
```